+++ tests/test.php
@@ -106,6 +106,8 @@ include_once("individualAccountRequest.php");
 include_once("individualAccountRequestLog.php");
 include_once("individualAccountAttachment.php");
 include_once("individualAccountAttachmentLog.php");
+include_once("businessAccountRequest.php");
+include_once("businessAccountRequestLog.php");
 include_once("pixInternalTransactionReport.php");
 include_once("pixInternalTransactionReportLog.php");